Reorder revert and reset
Introducing revert first suggests this is the "normal" or "better"
way to do things.
It also reduces the typing required, as we can then go on to delete
both the deliberately wrong commit and the revert commit, instead of
resetting first, then retyping the same example in order to have
something to reset.