Update theme and Makefile for new reveal.js version
The official way to do this is using the .scss files, but that
requires sass (recommended via npm install).
I previously edited css files, so that seems the quickest way
to get back to my previous position of a working slideshow with custom
styling.

Update Makefile so that I can rebuild during development,
by adding the Makefile and theme as dependencies.